requirements.txt
tensorflow>=2.10
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
pillow
opencv-python
